pyproject.toml
[project]
name = "itcobkai"
version = "0.1.0"
description = "サークルのOB会を大人数で行うために作成したボイスチャットです"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "livekit>=1.1.2",
    "livekit-agents>=1.4.4",
    "opencv-python>=4.13.0.92",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
